+++ Sources/HTMLParser/AST/HTMLDocument+Sequence.swift
@@ -1,0 +1,11 @@
+extension HTMLDocument: Sequence {
+    public func makeIterator() -> IndexingIterator<[HTMLNode]> {
+        children.makeIterator()
+    }
+}
+
+extension HTMLElement: Sequence {
+    public func makeIterator() -> IndexingIterator<[HTMLNode]> {
+        children.makeIterator()
+    }
+}